[level 0] Title: 숨어있는 숫자의 덧셈 (1), Time: 0.01 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/lv0/120851. 숨어있는 숫자의 덧셈 （1）/숨어있는 숫자의 덧셈 （1）.cpp
@@ -11,8 +11,7 @@ int solution(string my_string) {
         // my_string에서 자연수를 찾음
         if (e <= '9' && e >= '0')
         {
-            // 찾은 자연수를 char에서 int로 변환 후 연산
-            answer += static_cast<int>(e) - static_cast<int>('0');
+            answer += e - '0';
         }
     }
     return answer;
